Start of converting proto messages into cty.Value
The idea here is to allow a plugin for a hypothetical app which uses
HCL for configuration to return a message that the app can convert into
a cty.Value to include in EvalContext for further evaluation. It's not
yet complete enough to achieve that, but we at least have basic singleton
attribute value decoding here.